execution runner now plans the tasks which should allow for each stage to be handled as it starts rather than in advance.
