+++ orca-core/src/main/groovy/com/netflix/spinnaker/orca/pipeline/ExecutionLauncher.java
@@ -17,18 +17,10 @@
 package com.netflix.spinnaker.orca.pipeline;
 
 import java.io.IOException;
-import java.util.ArrayList;
-import java.util.Collection;
-import java.util.List;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.netflix.appinfo.InstanceInfo;
-import com.netflix.spinnaker.orca.pipeline.model.DefaultTask;
 import com.netflix.spinnaker.orca.pipeline.model.Execution;
-import com.netflix.spinnaker.orca.pipeline.model.Stage;
 import lombok.extern.slf4j.Slf4j;
-import static com.google.common.collect.Lists.reverse;
-import static com.netflix.spinnaker.orca.ExecutionStatus.NOT_STARTED;
-import static java.lang.String.format;
 
 @Slf4j
 public abstract class ExecutionLauncher<T extends Execution> {
@@ -36,16 +28,13 @@ public abstract class ExecutionLauncher<T extends Execution> {
   protected final ObjectMapper objectMapper;
   protected final InstanceInfo currentInstance;
   private final ExecutionRunner runner;
-  private final Collection<StageDefinitionBuilder> stageDefinitionBuilders;
 
   protected ExecutionLauncher(ObjectMapper objectMapper,
                               InstanceInfo currentInstance,
-                              ExecutionRunner runner,
-                              Collection<StageDefinitionBuilder> stageDefinitionBuilders) {
+                              ExecutionRunner runner) {
     this.objectMapper = objectMapper;
     this.currentInstance = currentInstance;
     this.runner = runner;
-    this.stageDefinitionBuilders = stageDefinitionBuilders;
   }
 
   public T start(String configJson) throws IOException {
@@ -53,53 +42,9 @@ public abstract class ExecutionLauncher<T extends Execution> {
     if (shouldQueue(execution)) {
       log.info("Queueing {}", execution.getId());
     } else {
-      planStages(execution);
       runner.start(execution);
     }
     return execution;
-  }
-
-  private void planStages(T execution) {
-    List<Stage<T>> stages = new ArrayList<>(execution.getStages()); // need to clone because we'll be modifying the list
-    stages.stream().forEach(this::planStage);
-  }
-
-  private void planStage(Stage<T> stage) {
-    StageDefinitionBuilder builder = findBuilderForStage(stage);
-    builder
-      .preStages()
-      .forEach(preStage -> {
-        List<Stage> stages = stage.getExecution().getStages();
-        int index = stages.indexOf(stage);
-        stages.add(index, preStage);
-        planStage((Stage<T>) preStage);
-      });
-    reverse(
-      builder
-        .postStages())
-      .forEach(postStage -> {
-        List<Stage> stages = stage.getExecution().getStages();
-        int index = stages.indexOf(stage);
-        stages.add(index + 1, postStage);
-        planStage((Stage<T>) postStage);
-      });
-    builder
-      .taskGraph()
-      .forEach(taskDef -> {
-        DefaultTask task = new DefaultTask();
-        task.setId(taskDef.getId());
-        task.setName(taskDef.getName());
-        task.setStatus(NOT_STARTED);
-        stage.getTasks().add(task);
-      });
-  }
-
-  private StageDefinitionBuilder findBuilderForStage(Stage<T> stage) {
-    return stageDefinitionBuilders
-      .stream()
-      .filter(builder1 -> builder1.getType().equals(stage.getType()))
-      .findFirst()
-      .orElseThrow(() -> new NoSuchStageDefinitionBuilder(stage.getType()));
   }
 
   protected abstract T parse(String configJson) throws IOException;
@@ -111,16 +56,6 @@ public abstract class ExecutionLauncher<T extends Execution> {
    */
   protected boolean shouldQueue(T execution) {
     return false;
-  }
-
-  public interface ExecutionRunner {
-    void start(Execution execution);
-  }
-
-  public static class NoSuchStageDefinitionBuilder extends RuntimeException {
-    public NoSuchStageDefinitionBuilder(String type) {
-      super(format("No StageDefinitionBuilder implementation for %s found", type));
-    }
   }
 }
